Migliorata la cancellazione delle iscrizioni

+++ models/registration.js
@@ -1,5 +1,4 @@
 const mongoose = require('mongoose');
-//const Exam = require('./exam');
 
 const registrationSchema = mongoose.Schema({
         idEsame: {
@@ -28,10 +27,13 @@ const registrationSchema = mongoose.Schema({
 
 registrationSchema.index({idEsame: 1, studente: 1, data: 1}, {unique: true});
 
-/*
 registrationSchema.statics.deleteRegistration = async (id) => {
-    return Registration.deleteOne({_id: id})
-};*/
+    const registration = await Registration.findOne({_id:id});
+    if(registration._id){
+        await mongoose.model('Exam').increaseFree(registration.idEsame);
+        await Registration.deleteOne({_id:id});
+    }
+};
 
 const Registration = mongoose.model('Registration', registrationSchema);
